don't run query if there are no results [touch:6431]{t:90}

+++ core/EE_Session.core.php
@@ -677,12 +677,14 @@ do_action( 'AHEE_log', __FILE__, __FUNCTION__, '' );/**
 				 $IN[] = "'" . str_replace( '_transient_timeout_', '', $expired_session ) . "'";
 			 }
 			 $IN = implode( ', ', $IN );
-			 $SQL = "
-			 	DELETE FROM {$wpdb->options}
-			 	WHERE option_name
-			 	IN ( $IN );
-			 ";
-			 $wpdb->query( $SQL );
+			 if ( ! empty( $IN )) {
+				 $SQL = "
+					DELETE FROM {$wpdb->options}
+					WHERE option_name
+					IN ( $IN );
+				 ";
+				 $wpdb->query( $SQL );
+			 }
 		 }
 	 }
